requirements.txt
# Core libraries
python-dotenv>=1.0.0
rich>=13.5.0
inquirer>=3.1.3
colorama>=0.4.6
pyyaml>=6.0.1

# API clients
dune-client>=1.1.0
tls-client>=0.2.2
fake-useragent>=1.1.3
requests>=2.31.0
aiohttp>=3.8.5
httpx>=0.24.1
websockets>=11.0.3
python-telegram-bot>=13.15
web3>=6.11.3
solana>=0.29.2
solders>=0.18.1

# Data processing
pandas>=2.0.0
numpy>=1.24.0
tabulate>=0.9.0
openpyxl>=3.1.0  # Required for Excel export functionality

# Terminal UI
prompt-toolkit>=3.0.36
curses-menu>=0.5.0
blessed>=1.20.0

# Utils
pydantic>=2.4.0
typing-extensions>=4.7.0
cryptography>=41.0.4
jsonschema>=4.17.3  # Required for config validation